add sanity checks we don't sign tx including dummy addr
Somewhat a follow-up to 649ce979abfeb18a285dec792d455ecf137c6382.

This adds some safety belts so we don't accidentally sign a tx that
contains a dummy address.
Specifically we check that tx does not contain output for dummy addr:
- in wallet.sign_transaction
- in network.broadcast_transaction

The second one is perhaps redundant, but I think it does not hurt.