Revert "qt: fix ReceiveWidget layout (fixes #9049)"
This reverts commit 21ffe321574f04373a673f4a1580806dfcc25758.

This layout is difficult to get right. Both the old and the new layout are somewhat buggy.
Reverting this only on the 4.5.x release branch. We can improve it on master.
The old layout has been out there in releases for many months, so it is safer to just use
that in the release branch for now.