batch payment manager:
The class TxBatcher handles the creation, broadcast and replacement
of replaceable transactions. Callers (LNWatcher, SwapManager) use
methods add_payment_output and add_sweep_info. Transactions
created by TxBatcher may combine sweeps and outgoing payments.

TxBatcher is always used when sweeping. When sending outgoing
payments with the GUI, its use is conditional to the config variable
WALLET_BATCH_RBF. This means that this config variable essentially
becomes a GUI setting.

Transactions created by TxBatcher will have their fee bumped
automatically (this was only the case for sweeps before).

Note that WALLET_BATCH_RBF must be enabled *before* the first
transaction is broadcast. It is no longer possible to add outputs
to a payment that was created with the option disabled.

    wallet:
     - instead of reading config variables, make_unsigned_transaction
       takes new parameters: base_tx, send_change_to_lighting

   tests:
     - add tests in test_txbatcher.py
     - remove test_sswaps.py
     - force regtests to use MPP, so that we sweep transactions with
       several HTLCs. This forces the payment manager to aggregate
       HTLC tx inputs.

   GUI:
    - if coin control is active, disable batching
    - forward submarine swaps can be batched

Note: If batching is enabled, the GUI will not merge duplicate
outputs. To enable that, we would need extra logic that matches
txbatcher._base_tx outputs to txbatcher.batch_payments
(Note that this is also what we would need in order to make payments
robust to reorgs, so it might be worth considering at one point.)