change plugins keyfile path to XDG spec
changes the plugins keyfile to either
XDG_CONFIG_HOME/electrum/plugins_key, or if empty to
HOME/.config/electrum/plugins_key or if HOME is not set either (which
should not happen in practice), to /etc/electrum/plugins_key.

Also adds a `bring_to_front()` call after installing external plugin so
the PluginsDialog is brought back in front.