implement release channels
implements release channels so the user can select either a stable or
beta release channel in the SettingsDialog for update notifications.
adapts the release_www.sh and make_download scripts to handle alpha
and beta versions which are detected by an 'a' or 'b' in the version name.
the version fields are added into the extradata field of the version file as
'version_alpha' and 'version_beta' and the extradata field is now being signed.
The update checker is supposed to return the newest version of those
contained in the version file which is
allowed according to the user configured option.